# ecommerce-site-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API",
        "description": "Documentation for E-Commerce API with Delivery Management",
        "contact": {
            "email": "admin@ecommerce.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        },
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/map/geocode": {
            "post": {
                "tags": [
                    "Map"
                ],
                "summary": "Geocode address",
                "description": "Convert text address to GPS coordinates",
                "operationId": "geocodeAddress",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "address"
                                ],
                                "properties": {
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main Street, Douala, Cameroun"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Geocoding successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "latitude": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "longitude": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "display_name": {
                                                    "type": "string"
                                                },
                                                "address_details": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Address not found"
                    }
                }
            }
        },
        "/api/map/route": {
            "post": {
                "tags": [
                    "Map"
                ],
                "summary": "Calculate route",
                "description": "Calculate optimal route between two points",
                "operationId": "calculateRoute",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "start",
                                    "end"
                                ],
                                "properties": {
                                    "start": {
                                        "properties": {
                                            "latitude": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "format": "float"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "end": {
                                        "properties": {
                                            "latitude": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "format": "float"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "profile": {
                                        "type": "string",
                                        "default": "driving",
                                        "enum": [
                                            "driving",
                                            "walking",
                                            "cycling"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Route calculated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "distance": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "distance_text": {
                                                    "type": "string"
                                                },
                                                "duration": {
                                                    "type": "integer"
                                                },
                                                "duration_text": {
                                                    "type": "string"
                                                },
                                                "geometry": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "number"
                                                        }
                                                    }
                                                },
                                                "steps": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/map/distance": {
            "post": {
                "tags": [
                    "Map"
                ],
                "summary": "Calculate distance",
                "description": "Calculate straight line distance between two points",
                "operationId": "calculateDistance",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "point1",
                                    "point2"
                                ],
                                "properties": {
                                    "point1": {
                                        "properties": {
                                            "latitude": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "format": "float"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "point2": {
                                        "properties": {
                                            "latitude": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "format": "float"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Distance calculated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "distance": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "distance_text": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}